Another run change

+++ models_mage_2dec.py
@@ -792,7 +792,7 @@ class MaskedGenerativeEncoderViT(nn.Module):
         attn=attn.clone().detach()
         # Latent another transformation?
         print(attn.shape)
-        attn_onehot = torch.nn.functional.one_hot(attn.argmax(2), num_classes=7)
+        attn_onehot = torch.nn.functional.one_hot(attn.argmax(2), num_classes=7).to(latent.dtype)
         print(attn_onehot.shape)
         print(attn_onehot.transpose(-1, -2).shape)
         print(latent.shape)
